requirements.txt
pytest~=7.1.2
selenium~=4.1.3
webdriver-manager~=3.5.4
python-dotenv~=0.20.0
allure-pytest~=2.9.45
faker~=13.7.0
pytest-xdist~=2.5.0
pytest-rerunfailures~=10.2
requests~=2.27.1
pydentic